refactor: UI와 비즈니스 로직 분리
- 계산 로직을 calculationService로 분리

- UI 유틸리티를 uiUtils로 분리

- Custom Hooks로 상태 관리 로직 분리

- 컴포넌트에서 비즈니스 로직 제거하여 관심사 분리
